* remove scala code * check style * support spark structured sql

+++ sylph-connectors/sylph-mysql/src/main/java/ideal/sylph/plugins/mysql/MysqlAsyncJoin.java
@@ -212,11 +212,11 @@ public class MysqlAsyncJoin
     {
         @Name("cache.max.number")
         @Description("this is max cache number")
-        private int maxNumber = 1000;
+        private long maxNumber = 1000;
 
         @Name("cache.expire.number")
         @Description("this is cache expire SECONDS")
-        private int cacheTime = 300;   // 5 minutes
+        private long cacheTime = 300;   // 5 minutes
 
         @Name("url")
         @Description("this is mysql jdbc url")
@@ -236,12 +236,12 @@ public class MysqlAsyncJoin
 
         public int getCacheTime()
         {
-            return cacheTime;
+            return (int) cacheTime;
         }
 
         public int getCacheMaxNumber()
         {
-            return maxNumber;
+            return (int) maxNumber;
         }
 
         public String getJdbcUrl()
